execution_core/requirements.txt
# Optional dependencies:
# - mysql-connector-python (only if you use --db-apply)
mysql-connector-python
